Updated stencil control to force update at start
Stencil now listen on monocle:load to update when
Monocle is loaded. Was not working previously. Was
working in the test, certainly due to the fact the
provided book's first component was 1 page long,
so the 2nd component was loading too. The loading
of the 2nd component was initiating the update.